+++ stac/src/item.rs
@@ -803,7 +803,8 @@ mod tests {
 
     #[test]
     fn deserialize_invalid_type_field() {
-        let mut item: Value = crate::io::read_json("data/simple-item.json").unwrap();
+        let mut item: Value =
+            serde_json::to_value(crate::read::<Item>("data/simple-item.json").unwrap()).unwrap();
         item["type"] = "Item".into(); // must be "Feature"
         assert!(serde_json::from_value::<Item>(item).is_err());
     }